Noticket: don't try to remove non existing user
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@24098 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/systemtests/src-test/org/ejbca/core/protocol/ws/CommonEjbcaWS.java
@@ -915,7 +915,9 @@ public abstract class CommonEjbcaWS extends CaTestCase {
             assertEquals("JurisdictionCountry=DE,JurisdictionState=Stockholm,JurisdictionLocality=Solna,CN=EVTLSEJBCAWSTEST", CertTools.getSubjectDN(cert));
         } finally {
             // Clean up immediately
-            endEntityManagementSession.deleteUser(admin, "EVTLSEJBCAWSTEST");
+            if (endEntityManagementSession.existsUser("EVTLSEJBCAWSTEST")) {
+                endEntityManagementSession.deleteUser(admin, "EVTLSEJBCAWSTEST");
+            }
             internalCertStoreSession.removeCertificate(CertTools.getFingerprintAsString(cert));
         }
     }
